requirements.txt
Flask==3.0.2
Flask-SQLAlchemy==3.1.1
pandas==2.2.1
scikit-learn==1.5.0  # 使用最新版本
mysql-connector-python==8.3.0